More work on #60.
Fix directed presence. Well, fixed in the sense that it works the way it used to.
After the bare JID <iq/> stuff is done, need to work on using the local process_stanza() for
outgoing stanzas instead of writing to their stream. Seems weird, but that will do all
the routing and such, which is a benefit of all this crap.