Implement an XML parser using expat.
This tries to use expat first, then falls back to REXML.
There is a major difference: REXML will detect not-well-formed XML, and still manage
to patch incomplete stanzas together okay. expat will never recognize a bad stanza, but
instead will wait for the rest of it to be valid or just sit there until the buffer explodes.

I guess this closes #46.